roll dynamic_mappings inside mappings
